Move other tokenizer tests
