+++ tests/ParadeDB.EntityFrameworkCore.IntegrationTests/QueryTests.cs
@@ -342,6 +342,35 @@ public sealed class MatchAllTests : TestBase
     }
 
     [Test]
+    public void ToString_EscapesOptionKeysAndStringValues()
+    {
+        var tokenizer = Tokenizer.Whitespace(new() { ["ali'as"] = "'escape me'", ["trim"] = true });
+
+        tokenizer.ToString().ShouldBe("pdb.whitespace('ali''as=''escape me''','trim=true')");
+    }
+
+    [Test]
+    public void ToString_EscapesRegexPattern()
+    {
+        var tokenizer = Tokenizer.RegexPattern("isn't", []);
+
+        tokenizer.ToString().ShouldBe("pdb.regex_pattern('isn''t')");
+    }
+
+    [Test]
+    public void ToString_ThrowsArgumentExceptionForUnsupportedOptionValue()
+    {
+        var tokenizer = Tokenizer.Whitespace(new() { ["bad"] = 1m });
+
+        var exception = Should.Throw<ArgumentException>(() => tokenizer.ToString());
+
+        exception.ParamName.ShouldBe("options");
+        exception.Message.ShouldContain(
+            "Tokenizer option 'bad' has unsupported value type 'Decimal'"
+        );
+    }
+
+    [Test]
     public async Task MatchAny()
     {
         await using var context = DbFixture.CreateContext();
